# specs/001-ai-rag-eval-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI RAG-Based Document Evaluation & QnA Platform API
  description: API for document upload, RAG-based QnA, and automated evaluation
  version: 0.1.0
  contact:
    name: ELP Gemini Team
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /docs:
    get:
      tags: [documents]
      summary: List user documents
      description: Get all documents owned by current user
      operationId: listDocuments
      security:
        - sessionAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [uploading, indexing, ready, failed, deleted]
          description: Filter by document status
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/DocumentResponse'
                  total:
                    type: integer
                    example: 15
        '401':
          $ref: '#/components/responses/Unauthorized'
  /docs/upload:
    post:
      tags: [documents]
      summary: Upload document
      description: Upload a PDF document for analysis
      operationId: uploadDocument
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file (max 50MB)
                title:
                  type: string
                  maxLength: 255
                  example: My Research Paper
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /docs/{document_id}:
    get:
      tags: [documents]
      summary: Get document details
      description: Retrieve specific document information
      operationId: getDocument
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [documents]
      summary: Delete document
      description: Delete a document (soft delete)
      operationId: deleteDocument
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Document deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Document deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Question & Answer ====================
...
  /admin/users:
    get:
      tags: [admin]
      summary: List all users
      description: Get list of platform users
      operationId: listAllUsers
      security:
        - sessionAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUserResponse'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{user_id}:
    get:
      tags: [admin]
      summary: Get user details
      description: Retrieve specific user information and activity
      operationId: getUserDetails
      security:
        - sessionAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: integer
      description: Document ID
  schemas:
...
    DocumentResponse:
      type: object
      properties:
        id:
          type: integer
          example: 42
        title:
          type: string
          example: My Research Paper
        status:
          type: string
          enum: [uploading, indexing, ready, failed, deleted]
          example: ready
        file_size:
          type: integer
          example: 2457600
        uploaded_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DocumentDetailResponse:
      allOf:
        - $ref: '#/components/schemas/DocumentResponse'
        - type: object
          properties:
            qa_count:
              type: integer
              example: 15
            evaluation_count:
              type: integer
              example: 2

    # ==================== QnA Schemas ====================
...
    QALogResponse:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        model_name:
          type: string
        created_at:
          type: string
          format: date-time

    # ==================== Evaluation Schemas ====================
    EvaluationRunResponse:
      type: object
      properties:
        id:
          type: integer
          example: 5
        status:
          type: string
          enum: [pending, running, success, failed]
          example: success
        template_id:
          type: integer
          example: 1
        model_name:
          type: string
          example: gemini-2.0-flash-thinking-exp
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
...
    AdminUserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        is_admin:
          type: boolean
        document_count:
          type: integer
        qna_count:
          type: integer
        created_at:
          type: string
          format: date-time
    AdminUserDetailResponse:
      allOf:
        - $ref: '#/components/schemas/AdminUserResponse'
        - type: object
          properties:
            documents:
              type: array
              items:
                $ref: '#/components/schemas/DocumentResponse'
            recent_qna:
              type: array
              items:
                $ref: '#/components/schemas/QALogResponse'
            recent_evaluations:
              type: array
              items:
                $ref: '#/components/schemas/EvaluationRunResponse'
...
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: Error description
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...